requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
botasaurus==3.0.9
Brotli==1.1.0
casefy==0.1.7
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller-fix==1.0.4
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
cryptography==41.0.7
dateutils==0.6.12
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
geonamescache==2.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kaitaistruct==0.10
lxml==4.9.3
MarkupSafe==2.1.3
orjson==3.9.10
outcome==1.3.0.post0
packaging==23.2
psutil==5.9.6
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pydivert==2.1.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-amazon-paapi==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
selenium==4.5.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.9.0
ujson==5.9.0
Unidecode==1.3.7
urllib3==1.26.18
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
zstandard==0.22.0
